Create Contribution based on Financial type
There might be better ways to achieve this, (not a developer) but it seems to do the trick. If the order has VAT (tax) the value passed to Civi's API is the Subtotal (Total with VAT - VAT = Subtotal) Of course Civi and Woocommerce have to have the same TAX assigned otherwise there will be a mismatch.